refactor - code deduplication
